shadow-dom/elements-and-dom-objects/shadowroot-object/shadowroot-attributes/test-007.html

<!DOCTYPE html>
<!--
Distributed under both the W3C Test Suite License [1] and the W3C
3-clause BSD License [2]. To contribute to a W3C Test Suite, see the
policies and contribution forms [3].

[1] http://www.w3.org/Consortium/Legal/2008/04-testsuite-license
[2] http://www.w3.org/Consortium/Legal/2008/03-bsd-license
[3] http://www.w3.org/2004/10/27-testcases
-->
<html>
<head>
<title>Shadow DOM Test: Confirm activeElement return null or element and is readonly</title>
<link rel="author" title="Sergey G. Grekhov" href="mailto:sgrekhov@unipro.ru">
<link rel="author" title="Masaya Iseki" href="mailto:iseki.m.aa@gmail.com">
<link rel="help" href="http://w3c.github.io/webcomponents/spec/shadow/#attributes">
<meta name="assert" content="ShadowRoot Object: confirm activeElement return null or element and is readonly.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../../../testcommon.js"></script>
<link rel="stylesheet" href="/resources/testharness.css">
</head>
<body>
<div id="log"></div>
<script>
test(unit(function (ctx) {

	var d = newRenderedHTMLDocument(ctx);

	var host = d.createElement('div');
	d.body.appendChild(host);
	var s = host.createShadowRoot();

    assert_readonly(s, 'activeElement', 'confirm activeElement is readonly');
    assert_equals(s.activeElement, null, 'confirm activeElement return null');

	var inp = d.createElement('input');
	s.appendChild(inp);

	inp.focus();

    assert_not_equals(s.activeElement, null, 'confirm activeElement return the currently focused element in the shadow tree');
    assert_equals(s.activeElement.tagName, 'INPUT', 'confirm activeElement return the currently focused element in the shadow tree');

}), 'Shadow DOM Test: Confirm activeElement return null or element and is readonly');
</script>
</body>
</html>
